+++ test/unit/mysql_db_admin/status.py
@@ -43,9 +43,8 @@ class Mail(object):
 
     Methods:
         __init__ -> Class initialization.
-        get_time -> Stub method holder for SlaveRep.get_time.
-        get_name -> Stub method holder for SlaveRep.get_name.
-        upd_slv_time -> Stub method holder for SlaveRep.upd_slv_time.
+        add_2_msg -> add_2_msg method.
+        send_mail -> send_mail method.
 
     """
 
@@ -59,7 +58,8 @@ class Mail(object):
 
         """
 
-        pass
+        self.lag_time = lag_time
+        self.data = None
 
     def add_2_msg(self, data):
 
@@ -70,6 +70,8 @@ class Mail(object):
         Arguments:
 
         """
+
+        self.data = data
 
         return True
 
@@ -94,6 +96,7 @@ class Server(object):
 
     Methods:
         __init__ -> Class initialization.
+        upd_srv_stat -> upd_srv_stat method.
 
     """
 
@@ -126,7 +129,7 @@ class Server(object):
 
         """
 
-        pass
+        return True
 
 
 class UnitTest(unittest.TestCase):
@@ -141,9 +144,11 @@ class UnitTest(unittest.TestCase):
         test_stdout_suppress -> Test with standard out being suppressed.
         test_mail -> Test with emailing out.
         test_file -> Test with writing to file.
+        test_append_to_file -> Test with appending to file.
         test_mongo -> Test with mongo connection.
         test_non_json -> Test with in non-JSON format.
         test_json -> Test with in JSON format.
+        test_flatten_json -> Test with flatten option for JSON format.
 
     """
 
@@ -162,6 +167,8 @@ class UnitTest(unittest.TestCase):
         self.args_array = {"-j": True, "-z": True}
         self.args_array2 = {}
         self.args_array3 = {"-j": True}
+        self.args_array4 = {"-j": True, "-a": True, "-z": True}
+        self.args_array5 = {"-j": True, "-f": True, "-z": True}
 
     def test_stdout(self):
 
@@ -227,6 +234,24 @@ class UnitTest(unittest.TestCase):
 
     @mock.patch("mysql_db_admin.mongo_libs.ins_doc")
     @mock.patch("mysql_db_admin.gen_libs.write_file")
+    def test_append_to_file(self, mock_write, mock_mongo):
+
+        """Function:  test_append_to_file
+
+        Description:  Test with appending to file.
+
+        Arguments:
+
+        """
+
+        mock_write.return_value = True
+        mock_mongo.return_value = True
+
+        self.assertFalse(mysql_db_admin.status(self.server, self.args_array4,
+                                               ofile="FileName"))
+
+    @mock.patch("mysql_db_admin.mongo_libs.ins_doc")
+    @mock.patch("mysql_db_admin.gen_libs.write_file")
     def test_mongo(self, mock_write, mock_mongo):
 
         """Function:  test_mongo
@@ -274,6 +299,23 @@ class UnitTest(unittest.TestCase):
         mock_mongo.return_value = True
 
         self.assertFalse(mysql_db_admin.status(self.server, self.args_array))
+
+    @mock.patch("mysql_db_admin.mongo_libs.ins_doc")
+    @mock.patch("mysql_db_admin.gen_libs.write_file")
+    def test_flatten_json(self, mock_write, mock_mongo):
+
+        """Function:  test_flatten_json
+
+        Description:  Test with flatten option for JSON format.
+
+        Arguments:
+
+        """
+
+        mock_write.return_value = True
+        mock_mongo.return_value = True
+
+        self.assertFalse(mysql_db_admin.status(self.server, self.args_array5))
 
 
 if __name__ == "__main__":